+++ src/Farmer/Arm/Dns.fs
@@ -47,7 +47,7 @@ module DnsRecords =
         interface IArmResource with
             member this.ResourceName = this.Name
             member this.JsonModel =
-                {| this.Type.ResourceType.Create(this.Zone + this.Name, dependsOn = [ this.Zone ]) with
+                {| this.Type.ResourceType.Create(this.Zone/this.Name, dependsOn = [ ResourceId.create this.Zone ]) with
                     properties = [
                         "TTL", box this.TTL