+++ src/Farmer/Arm/Network.fs
@@ -100,9 +100,9 @@ type VirtualNetworkGateway =
         member this.ResourceName = this.Name
         member this.JsonModel =
             let dependsOn = [
-                ArmExpression.resourceId(virtualNetworks, this.VirtualNetwork).Eval() |> ResourceName
+                ResourceId.create(virtualNetworks, this.VirtualNetwork)
                 for config in this.IpConfigs do
-                    ArmExpression.resourceId(publicIPAddresses, config.PublicIpName).Eval() |> ResourceName
+                    ResourceId.create(publicIPAddresses, config.PublicIpName)
             ]
 
             {| virtualNetworkGateways.Create(this.Name, this.Location, dependsOn, this.Tags) with
@@ -117,8 +117,8 @@ type VirtualNetworkGateway =
                                         | DynamicPrivateIp -> "Dynamic", null
                                         | StaticPrivateIp ip -> "Static", string ip
                                     {| privateIpAllocationMethod = allocationMethod; privateIpAddress = ip
-                                       publicIPAddress = {| id = ArmExpression.resourceId(publicIPAddresses, ipConfig.PublicIpName).Eval() |}
-                                       subnet = {| id = ArmExpression.resourceId(subnets, this.VirtualNetwork, ResourceName "GatewaySubnet").Eval() |}
+                                       publicIPAddress = {| id = ResourceId.create(publicIPAddresses, ipConfig.PublicIpName).Eval() |}
+                                       subnet = {| id = ResourceId.create(subnets, this.VirtualNetwork, ResourceName "GatewaySubnet").Eval() |}
                                     |}
                                 |})
                         sku =
@@ -175,9 +175,9 @@ type Connection =
       PeerId : string option
       AuthorizationKey : string option
       Tags: Map<string,string>  }
-    member private this.VNetGateway1ResourceId = ArmExpression.resourceId(virtualNetworkGateways, this.VirtualNetworkGateway1)
-    member private this.VNetGateway2ResourceId = this.VirtualNetworkGateway2 |> Option.map(fun gw -> ArmExpression.resourceId(virtualNetworkGateways, gw))
-    member private this.LocalNetworkGatewayResourceId = this.LocalNetworkGateway |> Option.map(fun lng -> ArmExpression.resourceId(localNetworkGateways, lng))
+    member private this.VNetGateway1ResourceId = ResourceId.create(virtualNetworkGateways, this.VirtualNetworkGateway1)
+    member private this.VNetGateway2ResourceId = this.VirtualNetworkGateway2 |> Option.map(fun gw -> ResourceId.create(virtualNetworkGateways, gw))
+    member private this.LocalNetworkGatewayResourceId = this.LocalNetworkGateway |> Option.map(fun lng -> ResourceId.create(localNetworkGateways, lng))
 
     interface IArmResource with
         member this.ResourceName = this.Name
@@ -185,7 +185,6 @@ type Connection =
             let dependsOn =
                 [ Some this.VNetGateway1ResourceId; this.VNetGateway2ResourceId; this.LocalNetworkGatewayResourceId ]
                 |> List.choose id
-                |> List.map(fun r -> r.Eval() |> ResourceName)
             {| connections.Create(this.Name, this.Location, dependsOn, this.Tags) with
                 properties =
                      {| authorizationKey = this.AuthorizationKey |> Option.toObj
@@ -217,9 +216,9 @@ type NetworkInterface =
         member this.ResourceName = this.Name
         member this.JsonModel =
             let dependsOn = [
-               this.VirtualNetwork
+               ResourceId.create this.VirtualNetwork
                for config in this.IpConfigs do
-                   config.PublicIpName
+                   ResourceId.create config.PublicIpName
             ]
             {| networkInterfaces.Create(this.Name, this.Location, dependsOn, this.Tags) with
                 properties =
@@ -229,8 +228,8 @@ type NetworkInterface =
                             {| name = sprintf "ipconfig%i" (index + 1)
                                properties =
                                 {| privateIPAllocationMethod = "Dynamic"
-                                   publicIPAddress = {| id = ArmExpression.resourceId(publicIPAddresses, ipConfig.PublicIpName).Eval() |}
-                                   subnet = {| id = ArmExpression.resourceId(subnets, this.VirtualNetwork, ipConfig.SubnetName).Eval() |}
+                                   publicIPAddress = {| id = ResourceId.create(publicIPAddresses, ipConfig.PublicIpName).Eval() |}
+                                   subnet = {| id = ResourceId.create(subnets, this.VirtualNetwork, ipConfig.SubnetName).Eval() |}
                                 |}
                             |})
                     |}
@@ -246,7 +245,7 @@ type NetworkProfile =
     interface IArmResource with
         member this.ResourceName = this.Name
         member this.JsonModel =
-            let dependsOn = [ ArmExpression.resourceId(virtualNetworks, this.VirtualNetwork).Eval() |> ResourceName ]
+            let dependsOn = [ ResourceId.create(virtualNetworks, this.VirtualNetwork) ]
             {| networkProfiles.Create(this.Name, this.Location, dependsOn, this.Tags) with
                 properties =
                     {| containerNetworkInterfaceConfigurations =
@@ -260,7 +259,7 @@ type NetworkProfile =
                                     {| name = sprintf "ipconfig%i" (index + 1)
                                        properties =
                                         {| subnet =
-                                            {| id = ArmExpression.resourceId(subnets, this.VirtualNetwork, ipConfig.SubnetName).Eval() |}
+                                            {| id = ResourceId.create(subnets, this.VirtualNetwork, ipConfig.SubnetName).Eval() |}
                                         |}
                                     |})
                                 |}