+++ src/Farmer/Builders/Builders.Search.fs
@@ -16,11 +16,11 @@ type SearchConfig =
     /// Gets an ARM expression for the admin key of the search instance.
     member this.AdminKey =
         let expr = sprintf "listAdminKeys('Microsoft.Search/searchServices/%s', '2015-08-19').primaryKey" this.Name.Value
-        ArmExpression.create(expr, this.Name)
+        ArmExpression.create(expr, (ResourceId.create this.Name))
     /// Gets an ARM expression for the query key of the search instance.
     member this.QueryKey =
         let expr = sprintf "listQueryKeys('Microsoft.Search/searchServices/%s', '2015-08-19').value[0].key" this.Name.Value
-        ArmExpression.create(expr, this.Name)
+        ArmExpression.create(expr, (ResourceId.create this.Name))
     interface IBuilder with
         member this.DependencyName = this.Name
         member this.BuildResources location = [