+++ RELEASE_NOTES.md
@@ -1,7 +1,10 @@
 Release Notes
 =============
+## 1.2.0-beta2
+* Internal changes to better capture full resource IDs
+
 ## 1.2.0-beta1
-* Bastion Hosts: Create bastion hosts for accessing resources on a virtual network.
+* Bastion Hosts: Create bastion hosts for accessing resources on a virtual network
 * Gateway: Add VPN Client configuration
 * Service Bus: Set message TTL
 * DNS Zones: Basic Azure DNS support